Moved session code into its own module, so we could use it in compat module without circular dependency issues
